+++ py/desitarget/io.py
@@ -3674,9 +3674,6 @@ def read_targets_in_hp(hpdirname, nside, pixlist, columns=None, header=False,
           grouped HEALPixels, as fewer files will need to be read.
         - If `mtl` is ``True`` then this is just a wrapper on
           read_mtl_in_hp().
-        - Will generally fail if the target files in a given `hpdirname`
-          are formatted differently BUT tries to allow a special DESI
-          case of an M31-program-like file and a standard DESI file.
     """
     # ADM if quick is True, use the quick-code.
     if quick:
@@ -3732,52 +3729,27 @@ def read_targets_in_hp(hpdirname, nside, pixlist, columns=None, header=False,
             targs, hdr = read_target_files(
                 infile, columns=columnscopy, header=True,
                 downsample=downsample, verbose=verbose)
-            targets.append(targs)
+            # ADM restrict immediately to the HEALPixels of interest.
+            # ADM This facilitates a check for HEALPixels that touch the
+            # ADM tiles of interest but that don't include any TARGETS
+            # ADM that touch the tiles of interest. This allows high-
+            # ADM separation programs such as the M31 and C-19 stream
+            # ADM BRIGHT1B programs to have different data models.
+            ii = is_in_hp(targs, nside, pixlist)
+            if np.any(ii):
+                targets.append(targs[ii])
         # ADM if targets is empty, return no targets.
         if len(targets) == 0:
             if header:
                 return notargs, nohdr
             else:
                 return notargs
-
-        try:
-            targets = np.concatenate(targets)
-        except TypeError:
-            # ADM special casing to allow mixing of some very specific
-            # ADM files with different formats, which are M31 target
-            # ADM files and SGC stream target files.
-            check = (
-                np.any([t.dtype.names[0] == "RELEASE" for t in targets]) &
-                np.any(np.array(["W1MPRO" in t.dtype.names for t in targets])) &
-                np.any(np.array(["RVS_FLAG" in t.dtype.names for t in targets])) &
-                np.any(np.array(["PANDAS_G" in t.dtype.names for t in targets]))
-            )
-            if check:
-                dtforce = [t.dtype.descr for t in targets
-                           if t.dtype.names[0] == "RELEASE"][0]
-                nomforce = [t.dtype.names for t in targets
-                            if t.dtype.names[0] == "RELEASE"][0]
-                newtargets = []
-                for t in targets:
-                    done = np.zeros(len(t), dtype=dtforce)
-                    for col in nomforce:
-                        done[col] = t[col]
-                    newtargets.append(done)
-                targets = np.concatenate(newtargets)
-            else:
-                msg = f"Mismatched data models for files in {hpdirname}"
-                log.critical(msg)
-                raise TypeError(msg)
-
+        targets = np.concatenate(targets)
     # ADM ...otherwise just read in the targets.
     else:
         targets, hdr = read_target_files(
             hpdirname, columns=columnscopy, header=True,
             downsample=downsample, verbose=verbose)
-
-    # ADM restrict the targets to the actual requested HEALPixels...
-    ii = is_in_hp(targets, nside, pixlist)
-    targets = targets[ii]
 
     # ADM ...and remove RA/Dec columns if we added them.
     if len(addedcols) > 0:
@@ -3834,9 +3806,6 @@ def read_targets_in_tiles_quick(hpdirname, tiles=None, columns=None,
           TARGETIDs from this function and that approach should be
           identical, although the output may be ordered differently.
         - Based on a suggestion from Anand Raichoor.
-        - Will generally fail if the target files in a given `hpdirname`
-          are formatted differently BUT tries to allow a special DESI
-          case of an M31-program-like file and a standard DESI file.
     """
     start = time()
     # ADM generator for the FITS files in the passed directory.
@@ -3892,8 +3861,13 @@ def read_targets_in_tiles_quick(hpdirname, tiles=None, columns=None,
         )
 
     # AR reading + concatenating.
+    # ADM adding the len(ii) > 0 as there's no need to read empty files.
+    # ADM But, also, allows high-separation programs such as the M31 and
+    # ADM C-19 stream BRIGHT1B programs to have different data models, as
+    # ADM one of these programs may touch a given HEALPixel/file without
+    # ADM any targets in that file touching a tile.
     targets = [fitsio.read(fn, rows=ii, columns=columns) for
-               fn, ii in zip(fns, iis)]
+               fn, ii in zip(fns, iis) if len(ii) > 0]
 
     # ADM if targets is empty, return no targets.
     if len(targets) == 0:
@@ -3905,34 +3879,7 @@ def read_targets_in_tiles_quick(hpdirname, tiles=None, columns=None,
         # ADM return a zero-length array with the correct data model.
         targets = np.zeros(0, dtype=targets.dtype)
     else:
-        try:
-            targets = np.concatenate(targets)
-        except TypeError:
-            # ADM special casing to allow mixing of some very specific
-            # ADM files with different formats, which are M31 target
-            # ADM files and SGC stream target files.
-            check = (
-                np.any([t.dtype.names[0] == "RELEASE" for t in targets]) &
-                np.any(np.array(["W1MPRO" in t.dtype.names for t in targets])) &
-                np.any(np.array(["RVS_FLAG" in t.dtype.names for t in targets])) &
-                np.any(np.array(["PANDAS_G" in t.dtype.names for t in targets]))
-            )
-            if check:
-                dtforce = [t.dtype.descr for t in targets
-                           if t.dtype.names[0] == "RELEASE"][0]
-                nomforce = [t.dtype.names for t in targets
-                            if t.dtype.names[0] == "RELEASE"][0]
-                newtargets = []
-                for t in targets:
-                    done = np.zeros(len(t), dtype=dtforce)
-                    for col in nomforce:
-                        done[col] = t[col]
-                    newtargets.append(done)
-                targets = np.concatenate(newtargets)
-            else:
-                msg = f"Mismatched data models for files in {hpdirname}"
-                log.critical(msg)
-                raise TypeError(msg)
+        targets = np.concatenate(targets)
 
     if header:
         return targets, hdr
@@ -4045,6 +3992,7 @@ def read_targets_in_quick(hpdirname, shape=None,
 
     # ADM determine the relevant HEALPixels for the file NSIDE.
     filepixlist = nside2nside(nside, filenside, pixlist)
+
     targets = []
     for pix in filepixlist:
         infile = formatter.format(pix)